+++ src/LangtailNode.ts
@@ -13,11 +13,10 @@ import { APIPromise } from "openai/core"
 
 import { userAgent } from "./userAgent"
 import { Stream } from "openai/streaming"
-import { OpenAiBodyType } from "./getOpenAIBody"
 
 export const baseURL = "https://proxy.langtail.com/v1"
 
-export interface ILangtailExtraProps extends OpenAiBodyType {
+export interface ILangtailExtraProps {
   doNotRecord?: boolean
   metadata?: Record<string, any>
 }